requirements.txt
beautifulsoup4==4.12.2
black==23.9.1
blinker==1.7.0
bs4==0.0.1
cachetools==5.3.2
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.2.0
click==8.1.7
cloudevents==1.10.1
deprecation==2.1.0
distlib==0.3.7
filelock==3.12.4
flake8==6.1.0
Flask==2.3.3
functions-framework==3.4.0
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-appengine-logging==1.3.2
google-cloud-audit-log==0.2.5
google-cloud-core==2.3.3
google-cloud-logging==3.8.0
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpc-google-iam-v1==0.12.7
grpcio==1.59.3
grpcio-status==1.59.3
gunicorn==20.1.0
identify==2.5.30
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
numpy==1.26.0
packaging==23.1
pandas==2.1.0
pathspec==0.11.2
platformdirs==3.11.0
pluggy==1.3.0
pre-commit==3.4.0
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pyflakes==3.1.0
pytest==7.4.2
pytest-mock==3.11.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-mock==1.11.0
rsa==4.9
six==1.16.0
soupsieve==2.5
tzdata==2023.3
urllib3==2.0.4
virtualenv==20.24.5
watchdog==3.0.0
Werkzeug==3.0.1
